+++ src/main/java/com/trego/dto/PreOrderDTO.java
@@ -2,6 +2,7 @@ package com.trego.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -11,7 +12,7 @@ public class PreOrderDTO {
     private long orderId;
     private long addressId;
     private String mobileNo;
-    private double totalCartValue;
+    private BigDecimal totalCartValue;
     private double amountToPay;
     private double discount;
     private List<CartDTO> carts;
